+++ {{cookiecutter.project_slug}}/config/settings/production.py
@@ -10,6 +10,12 @@ from sentry_sdk.integrations.celery import CeleryIntegration
 {% endif %}
 
 {% endif -%}
+{%- if cookiecutter.use_ldap_authentication == 'y' -%}
+import ldap
+
+from django_auth_ldap.config import LDAPGroupQuery, LDAPSearch, PosixGroupType
+
+{% endif -%}
 from .base import *  # noqa
 from .base import env
 
@@ -307,6 +313,58 @@ sentry_sdk.init(
 {% else %}
 sentry_sdk.init(dsn=SENTRY_DSN, integrations=[sentry_logging, DjangoIntegration()])
 {% endif -%}
+
 {% endif %}
+{%- if cookiecutter.use_ldap_authentication == 'y' -%}
+# AUTHENTICATION
+# ------------------------------------------------------------------------------
+# https://docs.djangoproject.com/en/dev/ref/settings/#authentication-backends
+AUTHENTICATION_BACKENDS = [
+    # https://django-auth-ldap.readthedocs.io/
+    "django_auth_ldap.backend.LDAPBackend"
+] + AUTHENTICATION_BACKENDS  # noqa
+
+# LDAP AUTHENTICATION
+# ------------------------------------------------------------------------------
+# https://django-auth-ldap.readthedocs.io/
+LDAP_BASE_DC = env("LDAP_BASE_DC")
+LDAP_BASE_OU = f"ou=groups,{LDAP_BASE_DC}"
+LDAP_BASE_GROUP = f"cn={env("LDAP_BASE_GROUP")},{LDAP_BASE_OU}"
+LDAP_PROJECT_GROUP = f"cn={{ cookiecutter.project_slug }},{LDAP_BASE_OU}"
+
+# Baseline configuration
+AUTH_LDAP_SERVER_URI = env("LDAP_SERVER_URI")
+AUTH_LDAP_BIND_DN = env("LDAP_BIND_DN", default="")
+AUTH_LDAP_BIND_PASSWORD = env("LDAP_BIND_PASSWORD", default="")
+AUTH_LDAP_USER_DN_TEMPLATE = "uid=%(user)s,ou=people," + LDAP_BASE_DC
+
+# Set up the basic group parameters
+AUTH_LDAP_GROUP_SEARCH = LDAPSearch(
+    LDAP_BASE_OU, ldap.SCOPE_SUBTREE, "(objectClass=posixGroup)"
+)
+AUTH_LDAP_GROUP_TYPE = PosixGroupType(name_attr="cn")
+
+# Simple group restrictions
+AUTH_LDAP_REQUIRE_GROUP = LDAPGroupQuery(LDAP_BASE_GROUP) | LDAPGroupQuery(
+    LDAP_PROJECT_GROUP
+)
+
+# Populate the Django user from the LDAP directory
+AUTH_LDAP_ALWAYS_UPDATE_USER = False
+
+AUTH_LDAP_USER_ATTR_MAP = {
+    "first_name": env("LDAP_FIRST_NAME_FIELD"),
+    "last_name": env("LDAP_LAST_NAME_FIELD"),
+    "email": env("LDAP_EMAIL_FIELD"),
+}
+
+AUTH_LDAP_USER_FLAGS_BY_GROUP = {
+    "is_active": LDAPGroupQuery(LDAP_BASE_GROUP) | LDAPGroupQuery(LDAP_PROJECT_GROUP)
+    "is_staff": LDAP_BASE_GROUP,
+    "is_superuser": LDAP_BASE_GROUP
+    # TODO: Wagtail group settings
+}
+
+{%- endif %}
 # Your stuff...
 # ------------------------------------------------------------------------------