+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All enhancements and patches to Cookiecutter Django will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2020-01-13]
+
+### Added
+
+- Add Django LDAP authentication integration (@jmiguelv)
+
 ## [2020-01-11]
 
 ### Changed